Add header fields to the header table on partial match
Previously we did not add header field to the header table if it is a
partial match against the current header table.  This commit changes
this scheme and adds header fields to the header table on partial
match.  hpack-test-cases data shows much better compressed result than
adding nothing.  We know from the experience that some header fields
are not effective and should not be indexed (e.g., :path), but this
commit does not incorporate that yet.