Implement table-based Huffman decoder.
This is heavily inspired by the work done by @tatsuhiro-t for nghttp2,
and ported to Python. In some micro-benchmarks focusing on decoding
primarily ASCII-text, the Huffman decoder in the form of this unrolled
state machine performs 4x faster than the naive Python version used
elsewhere in the HPACK module, decompressing a 1kB Huffman string in
~600usec compared to the naive version's 2.4msec.

This is without question a micro-optimisation, but given that the table
is not in flux anymore there's little advantage to using the naive
Huffman decoder.

Enormous gratitude is owed to @tatsuhiro-t for his work on this decoder.