fix encoding with generators
Previously we printed a debug message with the headers to encode.
This headers iterable could be a generator, which just debug-printed the
type information, but not the expected header values within.

As potential fix a previous commit simply converted it to a list - which
then rendered the generator empty and unusable. The tests didn't cover
this yet.

This commit removes the debug-print altogether, because each added
header is already debug-printed in the add() function. We add some
additional information to this existing debug print and remove the
top-level debug-print in the encode() function.

This commit adds a simple test case for passing a generator as headers
into encode().